+++ Utilities/gdcm/Source/MediaStorageAndFileFormat/gdcmSpacing.cxx
@@ -88,7 +88,8 @@ double frap(double frac[2], double startx, double maxden = 10 )
   const double error = startx - ((double) m[0][0] / (double) m[1][0]);
 
   /* now try other possibility */
-  ai = (maxden - m[1][1]) / m[1][0];
+  //!!! Is this supposed to be truncated to a long integer from a double?
+  ai = (long)((maxden - (double)m[1][1]) / (double)m[1][0]);
   m[0][0] = m[0][0] * ai + m[0][1];
   m[1][0] = m[1][0] * ai + m[1][1];
   //printf("%ld/%ld, error = %e\n", m[0][0], m[1][0],
